el: Depend on log module.
el library now emits logs using a logger (in particular, this can happen
in its shutdown), so we have to be sure the log module is loaded.

In order to add a relation between el and log modules, we have to ensure
the log module is registered in el_module_register. Since we can't rely on
the order of the constructors, call log_module_register manually, and make
it reentrant.

Change-Id: I18b1e63df43ce66cc79eb99271d403b84e4c0d14