el: fix ev_cache handling of reused-el.
Since the el library reuses the el as soon as they get destroyed, it's
possible that the qv of cache we generate from ev_cache_list contains ev
that has been destroyed and reused. In that case, we could call an ev
callback with the wrong set of arguments or when there is no good reason to
call it, potentially causing corruption or unexpected behaviors.

We now ensure all ev issued from ev_cache_list are tagged with a
generation that can be used to check if the ev has been reused.

This patch should have noticeable impacts only on recent branches since
the ev-types we use with ev_cache_list are limited to BEFORE, SIGNAL, IDLE
and PROXY, and only before get allocated-deallocated during the execution
of the callback (while the other ones are mostly static evs that remain
during the whole life of the process).

Change-Id: I8545895a89ca52ed97a7eaf4228f94bca1d943d6
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>