core-mem: Do not return NULL when allocating 0-size.
Problem:
~~~~~~~~
Some of our allocators (for example the mem-stack-pool, cf. 0e92542bf1)
return NULL when asking an allocation of size 0.

This is responsible for the bug #48066 (and maybe others) in case of an
IOP optional field whose type is an empty structure: we can't tell if it
is absent or present, because in both cases the pointer is NULL.

Solution:
~~~~~~~~~
We ensure in mp_*alloc functions that we don't return NULL when asking for
an allocation of size 0.

We use the fake pointer 0x1000 because it is in the range reserved by the
kernel, and should be aligned with everything.

Change-Id: I167d49f4163cb7ccce74e05103bcad1457199cd0
Refs: #48066
(cherry picked from commit 99ebbe55918c500cf5779ef67520891ee32ad8f2)