get rid of calls to readdir_r().
Recent glibcs have deprecated readdir_r() because this API is error-prone:
it requires allocating a buffer of the right size, which is not NAME_MAX
but fpatchconf(dir).

The suggested replacement is readdir() (which readdir_r() was supposed to
replace), which is safe-enough as long as you don't try to perform
concurrent readdir on the same DIR * object.

POSIX quote:
  The pointer returned by readdir() points to data which may be overwritten
  by another call to readdir() on the same directory stream. This data is
  not overwritten by another call to readdir() on a different directory
  stream.

Change-Id: Ia48ed5b82c150a2e275d374a2ea88ee378d7b341
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>