qhash: reduce the amount of generated functions.
A lot of accessors of the qh/qm is migrated to macros, reducing the list
of generated function to the needed ones.

As a consequence, we generate a lot less code at compilation time and thus
we can compile the repository faster. On my machine, with debug profile,
the compilation of the whole mmsx repository goes from 875s to 811s with
that patch (6% improvement).

Change-Id: I1aa599c15d111a1f2477ab66be037739e9fbb149
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>