qvector: reduce the amount of generated functions.
This is the same as the patch for the qhash, but this time for the
vectors. We only keep wipe/delete and the functions that use a block
(since the rewriter forces us to have functions for this).

This patch causes another compilation speed-up: from 811s to 752s. All in
all, the pair of patches provides a 14% decrease of compilation time on
mmsx.

Change-Id: I04a1709a500f0aa775019286f73f70bda9976713
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>