Build: add support for swift files.
We can now have source files written en Swift compiled in the same
libraries/binaries as C or C++ files.

The compilation places all the swift files for the same target in a single
swift module. As a consequence, all those modules can access their
internal symbols. Moreover, we do allow C code and Swift code to be
exposed into a single Swift module.

The compilation works mostly as-is: add your swift files to the source
list and they get compiled in the library/binary. A few variables can be
used to tweak the compilation behavior:

 {target}_SWIFTMAIN = whatever you want

This variable applies to binaries only and indicates that the main is
implemented in a swift source file, not in C.

 {target}_SWIFTMODULE = name

Allow overriding the name of the module used for the target. By default
the Swift module uses the same name as the target, with this parameter,
you can use another name. This is mostly useful in case the target name
contains a dash whish is forbidden in swift module names.

 {target}_SWIFTMIXED = whatever you want

Specifies that the same swift module contains both C and Swift code. This
is useful for libraries with both C and Swift for which you want to allow
access to the C code under the same module name as the Swift code.

Change-Id: I13ee253f18b2cb38da5524fe781873b07312aa3d
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>