swift: expose core.h API.
We add a module.map that specifies that the content of core.h can be
referenced as a module named common.core.

Note that this is a clang feature[1] that can be used to make dependencies
between modules more streamlined. However, this has the benefit of
allowing transparent access to C code from Swift code.

[1] http://clang.llvm.org/docs/Modules.html

Change-Id: Ica45caa6e93fe66995eeaf5e4443c25551d6c342
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>