doc: remove all the empty lines after the last element of the list
In order to be able to extend the list, we should not add any white line
after the end of the list.
This way, we are able to do something like

[bibliography]
include:<this file>
- extra element 1
- extra element 2

which is something that was forbidden before since the list contained a
white line and thus was not valid.

Change-Id: If842b926aeea068d820570b42497d4e35e0f64d0