ichannels: Fix fd leak when using ic_reply_fd_p.
This patch fixes a big regression of 1c9d3b792c: we are now leaking all
the fds given to ic_reply_fd_p.

Before 1c9d3b792c, the ic_msg_t was still the owner of the fd after
calling sendmsg, and the fd was properly closed when destroying the
message, which is not the case anymore.

Change-Id: Ifd566240feab463919290a720dbf615f4b3c8ab1