swift: add helper to wipe conforming C structures.
Since most of our structures come with a wipe() method, we can provide a
"standardized" approach to autowiping those structures.

Note that this is imperfect since nothing forbid the user to take a copy
of the wrapped object, which could have disastrous effects since those
objects must have a single owner (and must not be modified outside their
owning location).

Change-Id: Ic46e48f47dbe6190f9ccf82ac527e5a1643c2f38
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>