iopc: generate pointer constant to descriptors.
This is required for swift to access the descriptors by pointer (it's
mostly important for classes where we need the exact pointer to the
iop_struct_t to be in the object).

For safety reasons, Swift don't allow taking a pointer to a constant
value, so, if we want such pointers, we have to generate them ourselves in
the C code.

Change-Id: I98016ce4f70c14242eac2f21587518f6c5ee8512
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>