iopc: swift: generate structures and classes.
These IOP types are mapped to their corresponding Swift types (struct ->
struct, unless recursive, and class -> class) and respect the inheritance tree from the IOP
description (which is the main reason that makes me use separates types
from the C ones).

A C import/export interface is provided, it currently is not pretty
efficient, but we can imagine it will get faster later.

Change-Id: I94104b9844342951bf6fc66f43e3da37962ea8c1
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>