modules: allow to run a method during module initialization
During the initialization phase of a module, we just cleaned all the
module methods to rebuild them by taking care of the initializing module
dependencies. But, if the initializing module requires some others
modules, we won't load their module-methods because they should be
loaded as a dependency of the initializing module. However, the
initializing module has the state INITIALIZING... so we won't load its
module-methods neither its “dependent-of” module methods.

Hence the fix to load the “dependent-of” module methods of the
initializing module, crystal clear wasn't it ?

Change-Id: Iae4c7ce2ee4ad795f0be4438be4aa04284e2a88f