qps: make sure QPS chains sighandlers.
QPS supposed the sighandler was provided as a MODULE_PROVIDE(). However,
we have no proper code that do this before loading QPS, and as a
consequence, QPS-based programs (QKV, QRRD) didn't properly call the
standard panic handler, leading to the .debug file not being generated in
case of crash.

Following this patch, QPS module does not take a parameter anymore and
always suppose the handler to which it should chain is the previous
handler.

Change-Id: I6bf17701386e3d1b2a7c679919e7f5aa7fb30c95
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>