iop: fix: unpack/pack iop_tag correctly when it is an enum.
With --c-unions-use-enums, the enum size grows from uint16_t to int (or
possibly down to char, although not likely) so unpacking must be done with
the appropriate size or you may get garbage in your field.

We use the data offset when writing the data into the struct, and the data
offset + sizeof(a small enum) when reading the data from the struct.

Refs: #50352
Change-Id: I162331b71c24630937d1451a67b091af7656ed85