datetime: Stop caching result of lp_getsec.
lp_getsec() used a caching system in order to avoid calling time() at each
call. The value returned was the one evaluated when lp_gettv was called
the last time.

To ensure the freshness of this time, lp_gettv was called at the beginning
of the event loop, in el_timer_process.

Problems:
~~~~~~~~~
* if the event loop is long to process, lp_getsec could return timestamps
  a bit outdated.
* in threads, where there is no event loop, the result of lp_getsec() was
  only refreshed when lp_gettv was manually called! This could cause some
  timestamps to be *really* outdated!
  This probably explains the outdated logs coming from QTS that were
  observed on the SFR platform last week.

Solution:
~~~~~~~~~
Simply stop caching in lp_getsec(). According to a quick bench I've made,
calling time(NULL) directly is not that slow, and this does not justify to
cache its result (having correct values seems really more important).

For information, time(NULL) was between 16% and 60% slower than lp_getsec.

Change-Id: Ib5d2942e9fab8c87788fe93cef8f00de5f47b244