fix definition of __error__ when C11 is available
A regression was introduced in 9f5736121c91c9bdd37cac7f18375c974002845a
when dealing with the support of static asserts in C11. The __error__
needs to keep the same definition as with __GNUC__ defined, otherwise
all the code relying on __error__ to abort the compilation don't work
anymore (like the ASN.1 library).

Change-Id: Icc6ba8eef7a75303212f0fed87f16e9a17b4da11