core: add a 'if_assign()' macro.
This intend to provide a short-end for cases where we have to create a
scope based on the non-nullity of a pointer. Since is inspired by C++ that
allow the following construct:

 if (auto toto = dynamic_cast<Toto>(tata)) {
     /* Scope where toto is declared */
 }

The provided approach is quite similar:

 if_assign (toto, iop_obj_dynccast(toto, tata)) {
     /* Scope where toto is declared */
 }

The choice made here is to only support pointers types since it is the
most common use case where we need to use the returned value (in case of
integers, we mostly only want to make the distinction between positive and
negative values, which don't require a variable).

A while_assign with the same structure is provided as well.

Change-Id: I5d704183d157b6601ff904134124cf40de9fb8d7
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>