el: use enum type for flags.
Use the type of the enum that contains the flags instead of an int or an
unsigned to store flags. This may seem counter-intuitive since enum should
only describe a set of possible value, not a set of masks, but this has
several advantages to do so.

First, this documents the interface quite naturally since we now have a
direct reference to the type that contain the flags.

Secondly, this ensure the type has the proper size of signness, avoiding
issues in case someday we have more than 64 flags.

Thirdly, this play nicely with swift.

Change-Id: Ic96e712530811fce18f257f277bea2209feeef64
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>