el: use 64bit timeouts.
Currently we use 32bit signed timeout expressed in milliseconds. However, this is
quite restrictive since we overflow this integer in 24 days.

We recently fell twice in bugs where we tried to register a timer that run
in ~1 month, causing the timer to overflow, resulting in a negative
timeout. In one case this resulted in an infinite loop, in the other one
in a crash.

So, be safer and allow expressing such timers. We can now safely schedule
timers in the next 292 millions of years, which should be right enough for
the product (but still a bit short for the solar system which is expected
to die in ~5 billions of years).

Change-Id: I5491d0ffbd9ba7003264ed1f809244cb5aeb39e7
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>