el: fix corruption on el_child_unregister().
When we explicitly unregister an el_child on that is "already_exited", we
didn't properly remove it from the _G.before list, causing a corruption of
the list, and the firing of the content of the _G.ev_gc instead of the
content of _G.before in some cases.

We fix it by removing the complexity of the management of the ev_list: we
now ensure the list is always valid and thus we can also consider the
destruction of a el_t as being the move. As a consequence, all code fall
in the same code path, avoiding possible confusions.

Change-Id: Iaaf282bc722f994c418052159ba9268f8039a9e2
Closes: #47616
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>