iop: fix backward compat check on opt struct in newly added parent class
If a struct contains only optional fields, it is optional, even if marked
as "REQUIRED" in another struct.

If a new parent class is added to an existing class, and this parent
class contains such a field, it is backward compatible. Fix the code and
add a test for this particular case.

Change-Id: I2578ac4d2de6ed6e340b352eab5088c0b179406f