file-log: add file lock mechanism to compressed files
An issue was calling gzip multiple times for the same log file. This behavior
can occur when the log file is fairly big and that rotations are frequent.

This patch adds a locking mechanism to prevent the log file to compress logs
that are being compressed. Also, a rotation looks to delete old and big files
and a case can happen where a file being compressed can be deleted with the
original log file.

Thanks to this lock, we have the guarantee that compressions will ends
properly prior cleaning logs.

Closes: #43102
Closes: #47056
Change-Id: Iff71d2464d19192e9e6051797e30923633dfe4d7