python: http, allow to parse all responses
Before the python callback is only called when the http response is 200.
Some customers uses other http code like 500 to reply code fault. That
information is useful in case of the retry policy.

This patch updates the cb_parse_answer function with a new argument code,
to store the http code.

Change-Id: Ic42ab5029177f73871517b5c6d5afe6e2ec899c6
Closes: #51238