iop: add context to 'iop_for_each_field' callback
This patch adds the parameter 'stack' to the callback of the function
'iop_for_each_field'. This new parameter contains the data structure that
carries the context of the currently explored field.

Main advantage for now: when used for IOP struct validation or migration,
the errors can be like:

    "error with field `foo.bar[42].param`"

While 'iop_for_each_field' only allowed us to write errors like:

    "error with field `param`"

This implementation is based on a chained list embedded in the program
stack; it brings a CPU overhead of around +50% when compared to the
previous implementation.

It was tested about a qvector-based and a dlist-based implementations and
was slightly faster than the qvector version while the gap with the dlist
version was more significant. The bench material was some bigdata schema
samples from the product "One".

The old and faster implementation is kept under the name
'iop_for_each_field_fast' (same goes for the const variants).

A similar work was done for 'iop_for_each_st'.

Change-Id: Ia46ac2f5b8e7389d22d9b3e37c746abf714c7637