Revert "datetime: Stop caching result of lp_getsec."
This reverts commit c809ef1a8383634e9c77c8de0db1d00d161028bc.

The reason is that it seems there is some caching in the implementation of
time(), and as a result it can return a value strictly lower that the
tv_sec value of gettimeofday(), even if it's called later.

This strange behavior could cause subtle bugs in our products, so just
revert the patch in this stable branch. The following patch will
workaround the original issue solved by c809ef1a8 (which is the timestamps
broken in the threads).

Change-Id: I9a92f08ed0b8a21bec6b9b67a09288eeb5cd9055