Build: fix r.js build.
We didn't properly track the dependencies, causing the make www not to
call the r.js step when we modified a .js file. This one the consequence
of $(wildcard $e/**/*.json) not being recursive as expected. As a
consequence, only the first level of files was taken into account.

Moreover, we didn't properly set the dependencies for the minification,
causing the minifier not to run after r.js (a second run was needed).

Change-Id: I3528de808ebb581a0956725ad9efa8e43a90db1e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>