module: Enhance wording regarding methods registration.
The previous wording ("run") suggested this code was running methods,
while it actually only registers them for further calls (which is probably
inherited from the past, when the method callbacks were not cached, and
evaluated each time a method was called). Let's change the wording to make
it explicit.

Also remove a comment which talks about running method, while this is not
the case. It was more misleading than useful.

Change-Id: Ic6be294fc35efc37eb4e8412fc385334220ce8e8