ichannel: Don't consider canceled queries in ic_queue_len.
ic_queue_len is used in ic_is_empty in order to know if there are pending
queries on an ichannel.

Priori this patch, canceled queries (for example in case of timeout) were
accounting in this length, and as a consequence, an ichannel having only
timedout queries was not considered as empty. A very concrete impact is
that ctl stayed stuck even if the current query was canceled.

Even if we want to keep the canceled messages in the queries hash table
until we receive a response for them (because we want to keep the slot),
it makes no sense to consider they they are really in queue. So add a
dedicated counter for having the number of queries in queue (without the
canceled ones), and use it in ic_queue_len.

Change-Id: I42bf581df7345c5c5bacc104e9c57299fc8a2e3b
Refs: #51523