ic-rpc-http: Harmonize SOAP/REST error codes.
Before this patch, the REST API could return various error codes in case
of failure, not like the SOAP one which could just return
INTERNAL_SERVER_ERROR, as specified by the norm.

This was quite inconsistent, and could be surprising for clients, so
change the REST part in order to always use INTERNAL_SERVER_ERROR (and
document why we are doing that).

Change-Id: I5edc4802e8013a3a1bd663c2324b7f2e90f4b191
Closes: #51756