www: [Browserify] uglify during browserify.
In order to maintain source-maps correctly, uglify is now done during the
browserify action.

A new plug-in is then needed: uglifyify.

This avoid to generate source-maps once by browserify and then to
regenerate source-maps with uglify based on previous source-maps.
This last solution has sometimes lost of synchronization and we lost
source-maps for the final files.

Change-Id: I40846a9e168d62b18bc8a43384a2b22c83c22092
Closes: #51332