ic-rpc-http: Enhance SOAP interface error codes.
The previous patch was wrong since the clients could not distinguish an
IOP exception from the other errors since INTERNAL_SERVER_ERROR was always
thrown (and that nothing in the payload allowed to identify the different
cases.

So reserve INTERNAL_SERVER_ERROR for IOP exceptions, and use other codes
for the other cases.

Change-Id: I0688a471419d90ed20e09bdefac414a039864028
Refs: #51756