asn1/aper: switch from uint32_t to int32_t for native C type for enums
Since the native representation of C enum is 32bits signed integer, we
cannot really support enum values over INT32_MAX or not as C enumerations.

This change makes also the code consistent with what's done for BER (in
asn1-writer.c).

This patch unlocks the support of negative enumerated values.

Change-Id: Ie6830d0ae386fc443f241b5bf1b3b4bd05b3684c
Refs: #52024