asn1/aper: add safeguard for invalid enum values registration
When encoding an enumerated value in ASN.1 PER, we actually encode the
index of the value in the enumeration sorted in ascending order.

Example:

    foo ENUMERATED {
        A (14), -- Encoded as 0
        B (42), -- Encoded as 1
        C (99), -- Encoded as 3
        D (61), -- Encoded as 2
    }

Since we don't sort the array containing the values, the user that
registers an enum has to register it in ascending order or the encoding
will be broken.

This patch make sure that the enumerated values are registered in the
right order. It also ensures that there is no duplicated value.

Change-Id: I5860c6271e8e029c9e4827334b0b382ca84b02da
Refs: #52024