sort: implement dsort variants for signed integers
The dsort algorithms are meant to work with unsigned integers. In the
code, apply the transformation "f : x -> x - min" each time it is
necessary so the algorithms still works with unsigned integers even when
sorting signed integers.

The code is written in a way that forces gcc to generate the same code for
the already existing dsortX functions.

Avoid generating new "uniq_iX" functions: they don't need a different
implementation. The "uniq" algorithm doesn't really need the elements to
be sorted, it only need elements with the same values to be adjacent
(which of course can be considered as sorted with a non-standard sorting
order).

Change-Id: I5b64cd76346e21bafedde0dfa3a4bfac8a841c61