asn1/aper: add extension support in sequences
Only for "ComponentType" extensions (the ones with "..."):
"ExtensionAdditionGroup" extensions (the ones with "[[ <extensions> ]]")
are not needed for now.

Context:

In ASN.1, when specifying data, the specification author can prepare the
sequences for future additions by declaring extensions, this is done by
adding "..." at the end of the native fields (the root):

    Foo ::= SEQUENCE {
        i INTEGER (1..42),
        r REAL,
        ...
    }

Then, fields can be added in the extensions without breaking PER backward
compatibility:

    Foo ::= SEQUENCE {
        i INTEGER (1..42),
        r REAL,
        ...,
+       s OCTET STRING OPTIONAL,
+       bs BIT STRING OPTIONAL
    }

These fields should always be optional and they are a bit more expensive
to encode because we encode the length of each of them, allowing to skip
values that are not supported yet when reading the data with an outdated
grammar.

The full encoding for extended sequences is the following:

                                Length of the
                                extension bitmap
   Extension                    (the value encoded is l - 1 because
   bit                          the length cannot be null)
   ┌───┐                        ┌───────┐
   ┌───┬────────┬───────────────┬───────┬────────┬────────────────────┐
   │ 1 │ 001..0 │ <root fields> │ l - 1 │ 01..10 │ <extension fields> │
   └───┴────────┴───────────────┴───────┴────────┴────────────────────┘
       └────────┘                       └────────┘
       Bitmap for                       Bitmap for
       optional                         extension
       fields                           fields (length = l)
       (fixed length)

Change-Id: I627d5896e780ba6178328e23b61da84d447e9fc0
Close: #52301