asn1/aper: add default value for unknown extended values
This patch allows to define a default value to use when decoding an
extended enumerated value that is not registered.

Purpose: be able to skip non-critical enumerated fields when using an
outdated grammar.

Example:

Old grammar (G1):

 Foo ::= ENUMERATED {
     A (0),
     B (1),
     ...,
     C (2)
 }

New grammar (G2):

 Foo ::= ENUMERATED {
     A (0),
     B (1),
     ...,
     C (2),
     D (3),
     E (4)
 }

Before this patch, if we decoded a field foo of value D(3) with the
grammar (G1), we found an unregistered extension value and the decoding
failed. That could be a problem especially if we didn't really care
about the enumerated value.

With this patch, we can use a special value for unknown extension:

 foo.h:

 typedef enum foo {
     A,
     B,
     C,
     UNKNOWN_EXT,
 } foo_t;

 foo.c:

 ASN1_ENUM_BEGIN(foo);
     asn1_enum_reg_val(A);
     asn1_enum_reg_val(B);
     asn1_enum_reg_extension();
     asn1_enum_reg_val(C);
     asn1_enum_reg_ext_defval(UNKOWN_EXT);
 ASN1_ENUM_END();

This way, extended values D(3) and E(4), so as all other extended values
to come, will be encoded as "UNKNOWN_EXT".

Change-Id: I186eda964467f13630994e3696be420493934f02