Build: use sorcery to merge source maps.
Sorcery is a small utility that rebuild a sourcemap from the various
source maps that were generated during the build process in order to be
able to debug directly in the original sources, not an intermediary step.

However, it resolves all paths relatively to the path in which the map is,
so we have to write some hacky build rules in order to have it work
properly.

Change-Id: I859e7f656df97650ec9cd5659aceaa45fdd2f0c7
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>