Build: use iopc from lib-common instead of tools.
As a consequence, we must commit all the files included by the iopc.

A new mechanism is introduced in the build system in order to be able to
indicates a specific target should not be scheduled after the generated
headers/sources.

This should help checking the IOPC does not break anything.

Change-Id: Iff9cbeec859481e6d5581361371ab6e7a649f3e3
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>