iopc: Fix fields sorting in iopc_struct_sort_fields.
This function, introduced in e8d28c685, caches the current sort order so
that we can save some qv_sort when called again with the same sort order.

The problem is that it was not the only place fields order are modified:
iopc_struct_optimize also does that.

Take this into account by reseting the cache when optimizing a struct (and
vice-versa).

Change-Id: I15ceb4d2d5f577b7bd29a4f022d011e16d8737f2