iopc: add offset for void in union field.
Some code rely on the fact that all union fields have the same offset,
and optimize by using the offset of the first field for all of them.

This breaks when the first field is a void, which has no offset (0),
so replace the offset with the one of the first present field.

Change-Id: I2895c9b35583acfd1e1e022da46414a81b19f4ba