qps: Fix possible spool corruption at snapshot.
There was a big bug in the QPS snapshot code: if a QPS instance was
launched while a snapshotter fork was in progress (which can happen if the
previous instance crashed during a snapshot), this could corrupt the
spool. The exact issue is described in comment 7 of #52960.

To prevent this issue, the snapshotter process also takes a lock, so that
the main process can check if a snapshotter is in progress when it starts.
If it is the case, qps_open fails.

Change-Id: I45410898f79add689368dd8da0674b45d7bbbd8a
Closes: #52960