iop-dso: force fixed-up memory to be writable.
This will avoid crashes when using the old-fashioned fixing-up mechanism
to load DSOs.

This patch is needed because we switch to gold as the default linker, and
gold behavior is subtly different from the binutils default's, leading to
more memory to actually be read-only. The mprotect we add here removes a
protection :(

Change-Id: I1df437625b8eba8d54012a103c46da31020ce7e4
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>