iop-swift: don't add collection conformance to iop-arrays.
Instead of having direct conformance, add a .buffer property that is an
UnsafeBufferPointer which conforms to the collection protocol itself.

This dramatically reduce the amount of code emitted by the compiler since,
even if there is only a few function implemented in the IopArray protocol,
this causes all the default implementations of the Collection protocol to
be emitted for every single IOP array type.

We now emit a single function per array type + the initializers that
depend on the element type.

With this patch the P=release compile takes 13min instead of 28min on my
machine, and the generated binaries are smaller by 100MB.

Change-Id: I3c30225c848dbcddc434323a81075be95e4e89bb
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>