iopc: swift: emit IOP structs as classes.
Instead of having IOP structs in swift being represented by struct objects
conforming to the IopComplexType protocol, we now have a base class for
all of them. That base class conforms to IopComplexType and will as a
consequence contain all the instantiations of the methods of
IopComplexType avoiding those methods to be reinstantiated for every
single struct definition.

This reduces the size of the binaries of One on P=release by ~14MB and
reduces the compilation time by an additional minute.

Change-Id: I2fef99d51d2692f025d5b83af5b5851c101f04f5
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>