iopc: swift: limit the number of IopArray conformances.
Instead of adding an IopArray conformance for every single array, we now
do this only for arrays that are actually used: each package emit the
extensions for the __array_t type they use, instead of those of their
package.

In order to avoid duplicated extensions, we add a mechanism that check if
the deps of the current package already use that __array_t. If this is not
the case, we declare it ourself, otherwise, we use the existing extension.

This, however, is not a perfect solution since two package may use the
same array type without having a dependency on each other. This is
actually pretty rare (3 occurrences of the issue in mmsx). In order to
force the declaration in a location known by both packages in that case, a
new @swiftDumpArray attribute is introduced that force the creation of the
extension in the package of the definition of the type.

This patch avoids a lot of generated code. On One binaries, this is
another 15MB win (and 1min30 of compilation time too).

Change-Id: I37e31039164226d2ad8e7f5d0667ebdc598b6ba2
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>