thr-job: make thr_queue_sync reentrant.
When using thr_queue_sync, we expect the provided code to be executed on
the specified queue synchronously. However, if the queue is already
active on the current thread, this could cause a deadlock since the queue
would be waiting for itself.

thr_queue_sync now runs the job immediately if we already are on the
queue. This breaks the serial aspect of the queue but also ensure we can
queue stuff on a queue for itself.

New helpers are also provided to check if we are already running the
queue.

Change-Id: I3649d6d185c80edfb2e2477863393c5d6648b962
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>