thr-job: introduce a way to declare thread datas.
Currently, when running a job on the thread jobs, we often want to
accumulate the results using per-thread accumulators. To do this, we
allocate thr_parallelism_g accumulators before starting the jobs, and
deallocate them after the jobs have finished running. Each job can then
finds the accumulator associated to the thread it runs on by using the
thr_id().

This however is not suitable in case the list of threads is dynamic.
That's why we introduce a new mean to achieve this. The accumulators are
now handled by the thr_syn_t and are dynamically allocated. They are not
really associated to a thread, but stored in a stack. A job that wants to
use one accumulator must acquire it. If the stack is empty, a new one is
allocated, otherwise the first entry of the stack is popped.

When the job has finished using the accumulator, it pushes it back on the
stack by releasing it (otherwise there will be a memory leak).

Finally, when everything is done, the result can be collected from the
various allocated accumulators.

NOTE: in the future we could imagine to make the API much more 'reduce'
like by providing an allocator, a merger and a wiper. That way, we could
run the merging in parallel too.

Change-Id: I6bc01d8712887cd816e80e5e4bdffc5dc23f23fb
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>