thr-jobs: fork new threads to replaced blocked ones.
This patch introduces two things: first a new API to declare that a thread
enters/exit a blocking system call, secondly a mechanism to fork/kill
threads when the parallelism changes.

The thr-jobs library now keeps track of a "target_threads_count" that is
the number of threads it needs to achieve the thr_parallelism_g while
taking into account the fact that some threads may be blocked in system
calls.

That count is updated using the two calls thr_enter_blocking_syscall() and
thr_exit_blocking_syscall().

The thread-forking mechanism works as follow: threads are assigned ids.
When the target_threads_count grows, we fork threads up to the id
target_threads_count. Then, the threads themselves are responsible for
their own deactivation when the target_threads_count goes down. That check
is part of the run loop of the threads. When the target_threads_count goes
down, threads whose id is greater than that value stop consuming new jobs,
and goes in a waiting loop. If the target count does not goes up again in
the 5sec, the threads exits.

When a thread dies, the context of the thread is preserved (because it may
still contains some jobs to be run). A thread with the same id can be
forked again later if needed.

The thr-jobs library always keeps an idle thread that is ready to take on
new jobs when the target_threads_count grows.

Note that this patch includes some e_trace(0) that will be improved in a
future patch.

Change-Id: I6c39d7595b85b806d86f6aa73d013b15be0d9c26
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>