thr-jobs: thr_syn_wait() does not run jobs anymore.
In order to guarantee the forward advance of the execution, thr_syn_wait()
was implemented as a sub-run-loop of the thread-jobs. However, this causes
some jobs to be callable in the same thread where you are calling a
thr_syn_wait() in the middle of the code of a function. This adds a lot of
corner cases such as thread-local variable being allowed to be affected by
another job, which basically forbids their use in jobs.

In order to fix this, a thr_syn_wait() is now considered a blocking
syscall if it has to wait more than 100ms.

There is still a special handling of the main queue if we are on thread 0,
causing jobs of the main queue to be able to run if we call thr_syn_wait
on the main thread.

Change-Id: I2061b7f250a3d4f2b0669b93290c737241974852
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>