Build: disable dwarf compression when objcopy don't support it.
Because swift requires the gold linker and the gold linker is more modern
than the standard one on centos 6, the compression of the debug sections
get activated on Centos 6. However, gdb on that platform does not support
compressed debug sections leading to unexploitable backtraces.

This patch enables compression only when both the linker and objcopy
support debug-section compression, avoiding that caveat.

Change-Id: I53819810d187c5e078dd3e7b3ebded6ca519f3f8
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>