Build: www: rework browserify build.
This patches get rid of deamdify since it seems to conflict with
browserify-shim as soon as there is not minification between the two. That
way, it will be easier to disable minification in debug mode in a future
patch.

Moreover, uglification is now a separate pass that is performed on the
output of browserify.

Change-Id: Id9776eff0e230e6036691718c3e5c90deeace1cf
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>