iopc: Fix possible invalid read of iopc_typer_g.resolve_list.
When the resolution (first or second passed) failed, the
iopc_typer_g.resolve_list vector was left untouched, and its elements were
reused when compiling the next IOP file.

This never happens in the standalone iopc binary, but could easily cause
crashes when using the iopc library (for example in one-core, when
updating the schema).

Change-Id: Ice9ccfe90990c0e7e58a822eff1a6f9b1d66491a