Build: run jshint/tslint when copying the files.
When dealing with bundled-files, we now handle the linting of the
javascript and typescript sources when the files is copied into the .build
directory instead of waiting of doing all the files at once at the end.
That way, we only lint the files that were changed and we benefit from the
parallelisation of the tasks.

This can be disabled using the "standard" NOCHECK argument.

Change-Id: I183d30c5ecfed0a933e23ac8258c2a1604336f44
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>