container-qhash: forbid use of qm_get_def with incompatible default value
There can be an issue that is quite hard to catch when reviewing /
debugging triggered with qm_get_def that can get messy when using a
negative default value while the value type of the qm is unsigned.

In order to solve this issue and to ensure it cannot happen, this patch
ensures the default value provided to qm_get_def is compatible with the
value type of the qm.

Automated tests about qm_get_def are also added.

Change-Id: Ia5ed8177983886fc03ec119e142f702020b9dfdc