Build: www: ignore missing packages.
This is really a tricky one here: we don't want browserify to warn about
missing packages because we do have multiple bundles.

We could use the -x flag to mark a module as being in another bundle, but
this would be much harder (need to list all dependencies). So we simply
say browserify not to warn about external deps.

This however implies that we won't be able to easily spot missing packages
at compilation-time.

Change-Id: I825c3783b7a5ef2cd05a6dffd499ef3352f1b4b4
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>