Build: www: fix support of multi-bundle with typescript.
We now have a single place were we centralize the .d.ts imported or
generated at the root of the build directory. That way, it is in the
resolution path for all the libraries.

tslint cannot be used with typecheck rules anymore because it cannot
resolve those .d.ts properly.

Also, don't generate the .ts.d anymore since typescript can have circular
deps, this caused a lot of compilation warning for no actual improvement
in dep checking.

Change-Id: I39fc5c770f42fe5b521fed37ae825eb8e14d99a2
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>