python-common: auto release pylogger through registry.
Usually loggers are wiped automatically by Python when the variable is
not referenced. Unfortunately, when the logger is caught in a circular
dependency, it is not wiped until Python GC is run, and Python GC is
never run at the end of the program. Non-static loggers that are not
wiped at the end of the program generate a panic.

So use a global registry to clean pyloggers when modules are released
before the end of the program.

Change-Id: I2cafaa06b6fb81816224bb42c5dfe240d6b311f8
Refs: #53973